Fixed a bug and caching the inner loop
The bug actually does not affect the benchmark because we are dealing with
square matrices. Caching seems to improve the speed for small n, but strangely
not for large. Brentp said that v2 should be 10-15% faster. This also depends
on n. For small n, v1 is faster; for large n, v2 is faster.