Go: use ReadSlice in dict_v1.
This is uglier and not as general as ReadString, but ~28% faster.

It is faster because ReadString currently just calls ReadBytes,
which makes copy of a buffer, and then ReadString make a copy
again when converting to string. ReadSlice is limited to defaultBufSize
(4096), and reuses the buffer (but we make a copy of it when converting
to string).